adding refresh flow and abstracting oauth2 server into server handling pattern
